django/tastypie_bpmappers/myproject/requirements.txt
Django==1.5.4
bpmappers==0.7
django-tastypie==0.10.0
mimeparse==0.1.3
python-dateutil==2.1
six==1.4.1
